+++ Clients/haskell/king-game-client/king-game-client.cabal
@@ -27,13 +27,17 @@ source-repository head
 
 library
   exposed-modules:
-      Lib
+      KingClient
   other-modules:
       Paths_king_game_client
   hs-source-dirs:
       src
   build-depends:
       base >=4.7 && <5
+    , bytestring
+    , split
+    , json
+    , zeromq4-haskell
   default-language: Haskell2010
 
 executable king-game-client-exe
@@ -46,11 +50,6 @@ executable king-game-client-exe
   build-depends:
       base >=4.7 && <5
     , king-game-client
-    , zeromq4-haskell
-    , bytestring
-    , json
-    , mtl
-    , split
   default-language: Haskell2010
 
 test-suite king-game-client-test
