--- Clients/haskell/king-game-client/src/Lib.hs
@@ -1,6 +1,0 @@
-module Lib
-    ( someFunc
-    ) where
-
-someFunc :: IO ()
-someFunc = putStrLn "someFunc"
